refactor(theme): flatten theme directory structure
