+++ include/lvgl/lvgl.h
@@ -171,10 +171,10 @@
 #include "stdlib/lv_mem.h"
 #include "stdlib/lv_sprintf.h"
 #include "stdlib/lv_string.h"
-#include "themes/default/lv_theme_default.h"
 #include "themes/lv_theme.h"
-#include "themes/mono/lv_theme_mono.h"
-#include "themes/simple/lv_theme_simple.h"
+#include "themes/lv_theme_default.h"
+#include "themes/lv_theme_mono.h"
+#include "themes/lv_theme_simple.h"
 #include "tick/lv_tick.h"
 #include "widgets/lv_3dtexture.h"
 #include "widgets/lv_animimage.h"
